# rumicoach-backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: rumi.coach API
  version: 1.0.0
servers:
  - url: /v1
paths:
...
  /sessions:
    get:
      summary: List user's sessions
      tags:
        - Users
      operationId: GetSessions
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSessionPaginatedResponse'
        '401':
          description: Unauthorized
...
  /webhooks/telegram/{secret}:
    post:
      summary: Telegram incoming updates webhook
      tags:
        - Webhooks
      operationId: PostWebhooksTelegramSecret
      parameters:
        - name: secret
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /webhooks/revenuecat:
    post:
      summary: RevenueCat purchase events webhook (credits the minutes balance)
      tags:
        - Webhooks
      operationId: PostWebhooksRevenuecat
      responses:
        '200':
          description: Event processed (or deliberately ignored/duplicate)
        '401':
          description: Missing or wrong Authorization header
        '500':
          description: Transient failure — RevenueCat should retry
        '503':
          description: REVENUECAT_WEBHOOK_AUTH not configured on this deployment
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
...
    CommunicationSession:
      type: object
      properties:
        id:
          type: string
        startTime:
          type: string
          format: date-time
        duration:
          type: integer
        sessionType:
          type: string
        recapTitle:
          type: string
          nullable: true
          description: "A few words naming what the session was about, in the user's language. Use it as the row heading in the sessions list; fall back to the session-type label when absent."
        recap:
          type: string
          nullable: true
          description: "Short user-facing summary (2-3 sentences) of what the session covered and what the user took from it, in their language. Generated after the session ends, so it is absent for sessions still in progress, for very short ones, and always for the onboarding intro."
        userSessionInsight:
          type: string
          nullable: true
        userEvaluation:
          type: number
          format: float
          nullable: true
        userFeedback:
          type: string
          nullable: true
...
    UserSessionPaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CommunicationSession'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...